refactor: Tri cell context, cell renderer, bullet button.
