--- packages/zodspy/components/tri/tana-themes.tsx
@@ -1,1 +1,0 @@
-export const TanaThemes = () => {};
