+++ packages/zodspy/components/tri/tri-icons.tsx
